First version of the log-in and sign-up part.
I have completed a first version of the log-in and sign-up part
of the app. The app is connected to Firebase. It is possible to
create, delete and remove a user. There are three areas where this could
still improve a lot: error handling (especially that related to loss
of internet connection), allow users to recover their password or
confirm their email, and stop hardcoding strings.