HomeActivity V.05 Swipe and Refresh
HomeActivity is already provided with all of its basic functionallity,
included swipe and refresh on fragments. Some other minor changes in
this version are: renamed Test class to Activity and created class
Template.
There are some other minor and aesthetic changes to be made on
HomeActiviyt, but I'm not dealing with them for now. Instead I will
start with programming new activities.