Location tracker working.
There are several new thing included in this commit. On the one hand,
there is a first approach to the map configuration. On the other hand
there is a foreground service implemented which tracks the user
location. Both things should be close to their final version. But also,
there are also some experimental additions regarding to geofencing and
notifications. Those are all commented, since they are not polish and
there is high risk that they can compromise the good working of the rest
of the app.