Quick issues fixed and much more.
Some little more issues have been fixed here. But also, for mistake, a
lot more has been done related to Participations. This should be in some
different branch. But let's leave it that way. As has been said, the
concept Participation has been introduced to the model. When the user
enrolls an activity, a Participation object is created. Also there has
been created a new Activity that contains information about the on going
activity.