First version of participations completed
Here there is a first approach to all the issues related with the user
taking part in an activity, unless that related to the fact that he/she
shouldn't be allowd to be participating in more that one activity at
once. Also, lots of the checkings are being done in the client. This
should be done in Firebase. Should be fixed later.