Activities have already beacons.
Now it is posible to take part in an activity and events will happen
when you reach a beacon. The app will let you know and also will update
the information in Firestore. Finally, if the beacon you just reached is
the last one, it finishes the activity and the foreground service.