Trying to solve problems with notifications.
In this commit there is not actually much new. Only tryings to make the
notification vibrate, sound, light, full screen... But not being successfull.
Changes commited just in case the lines comented and the permissions in
the Manifest are usefull some time. The exact same applies to making the
back button after a notification work. Tons of alternatives have been
try but nothing seems to work.