Goal beacon included.
In this commit it has been introduced the concept of the goal beacon.
Also it has been refactored all the code that contains the logic of the
game in the service. Now it's much more readable and reliable. Still not
done the part related to the notifications (both regular notifications
and finish notifications). Also, there are some tasks that should be
done server side.