Included a real template of activity.
It has been included a real template of an activity of the 'Educativa
Naranja' type. It still lacks the educative content of the beacons. But
it contains already the beacons and the map. Also somer minor changes
has been made to the user interface, and the model of Map has been
included.