Created an activity to display beacon's challenges
There has been created an activity to display the beacon's challenges.
In that activity it has been included a picture that the user can zoom
in and out. Also it has been created the Fragment that allows to give an
answer when the activity is of the NARANJA type. The logic for that,
however, is not yet implemented.