It is already posible to display the track
A participant can watch his track. Also the foreground service updates
the locations. Some problems not solved yet at this point are the
following: we should check that the locations are good enough,
otherwise, the track is going to be rare. We should only show the track
to the participant once he has finished. We should allow the organizer
to see the track too.
Finally, if something is failing and we want to remove it, we just need
to comment a line in LocationService to stop updating locations. And
also we would only need to disable the buttons that lead to the track
activity.