New log-in page created.
The new log-in page has been created. Also, now it is possible for the
user to reset his password. Careful, because both that feature and the
one about e-mail verification are comented (disabled) in this version to
make testing easier. Moreover some other quick-fixes have been added.
Now it is possible for a participant to cancel his subscription to an
activity.